# document/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Content Manager Service",
        "contact": {},
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local testing"
        },
        {
            "url": "https://alinlandev.com/api/services/content",
            "description": "Server testing"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "List",
                "operationId": "GetList2",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "target-user",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "style": "simple",
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "example": 1
                        }
                    },
                    {
                        "name": "actor",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "style": "simple",
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "headers": {}
                    }
                },
                "deprecated": false
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create",
                "operationId": "PostCreate2",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "actor",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "style": "simple",
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "example": 1
                        }
                    },
                    {
                        "name": "target-user",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "style": "simple",
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRequest2"
                            },
                            "example": {
                                "type": "simple",
                                "name": "name",
                                "sku": "sku",
                                "description": "",
                                "status": 0,
                                "brand_id": 308,
                                "url_key": "",
                                "tax_class": 3,
                                "visibility": 2,
                                "barcode": "",
                                "dimensions": "",
                                "meta_title": "",
                                "meta_keyword": "",
                                "meta_description": ""
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "",
                        "headers": {}
                    }
                },
                "deprecated": false
            }
        },
...
        "/products/{productId}/category/assign": {
            "parameters": [
                {
                    "in": "path",
                    "name": "productId",
                    "required": true,
                    "schema": {
                        "type": "integer",
                        "minimum": 1
                    }
                }
            ],
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Assign categories",
                "operationId": "assignCategories",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "description": "",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_ids": {
                                        "type": "array"
                                    }
                                }
                            },
                            "example": {
                                "category_ids": [
                                    204,
                                    97
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "",
                        "headers": {}
                    }
                },
                "deprecated": false
            }
        },
        "/products/{productId}/category/unassign": {
            "parameters": [
                {
                    "in": "path",
                    "name": "productId",
                    "required": true,
                    "schema": {
                        "type": "integer",
                        "minimum": 1
                    }
                }
            ],
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Unssign categories",
                "operationId": "unassignCategories",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "description": "",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_ids": {
                                        "type": "array"
                                    }
                                }
                            },
                            "example": {
                                "category_ids": [
                                    204,
                                    97
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "",
                        "headers": {}
                    }
                },
                "deprecated": false
            }
        },
        "/products/{productId}/configurable": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Add Configurable",
                "operationId": "AddConfigurable",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "productId",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddConfigurableRequest"
                            },
                            "example": {
                                "product_ids": [
                                    6848,
                                    6869
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "",
                        "headers": {}
                    }
                },
                "deprecated": false
            }
        },
        "/products/{productId}/configurable/remove": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Remove Configurable",
                "operationId": "RemoveConfigurable",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "productId",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RemoveConfigurableRequest"
                            },
                            "example": {
                                "product_ids": [
                                    6848,
                                    6869
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "",
                        "headers": {}
                    }
                },
                "deprecated": false
            }
        },
...
        "/products/{productId}/related": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Add Related",
                "operationId": "AddRelated",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "productId",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_ids": {
                                        "type": "array"
                                    }
                                }
                            },
                            "example": {
                                "product_ids": [
                                    6848,
                                    6869
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "",
                        "headers": {}
                    }
                },
                "deprecated": false
            }
        },
        "/products/{productId}/related/remove": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Remove Related",
                "operationId": "RemoveRelated",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "productId",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_ids": {
                                        "type": "array"
                                    }
                                }
                            },
                            "example": {
                                "product_ids": [
                                    6848,
                                    6869
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "",
                        "headers": {}
                    }
                },
                "deprecated": false
            }
        },
...
        "/categories/{id}/products": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Products",
                "operationId": "Products",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "headers": {}
                    }
                },
                "deprecated": false
            }
        },
...
        "/attribute-sets/{id}/assign": {
            "post": {
                "tags": [
                    "Attribute Set"
                ],
                "summary": "Assign Attribute",
                "operationId": "AssignAttribute",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AssignAttributeRequest"
                            },
                            "example": {
                                "attribute_ids": [
                                    1
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "",
                        "headers": {}
                    }
                },
                "deprecated": false
            }
        },
        "/attribute-sets/{id}/unassign": {
            "post": {
                "tags": [
                    "Attribute Set"
                ],
                "summary": "Unassign Attribute",
                "operationId": "UnassignAttribute",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UnassignAttributeRequest"
                            },
                            "example": {
                                "attribute_ids": [
                                    1
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "204": {
                        "description": "",
                        "headers": {}
                    }
                },
                "deprecated": false
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateRequest2": {
                "title": "CreateRequest2",
                "required": [
                    "type",
                    "name",
                    "sku",
                    "description",
                    "status",
                    "brand_id",
                    "url_key",
                    "tax_class",
                    "visibility",
                    "barcode",
                    "dimensions",
                    "meta_title",
                    "meta_keyword",
                    "meta_description"
                ],
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "sku": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "brand_id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "url_key": {
                        "type": "string"
                    },
                    "tax_class": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "visibility": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "barcode": {
                        "type": "string"
                    },
                    "dimensions": {
                        "type": "string"
                    },
                    "meta_title": {
                        "type": "string"
                    },
                    "meta_keyword": {
                        "type": "string"
                    },
                    "meta_description": {
                        "type": "string"
                    }
                },
                "example": {
                    "type": "simple",
                    "name": "name",
                    "sku": "sku",
                    "description": "",
                    "status": 0,
                    "brand_id": 308,
                    "url_key": "",
                    "tax_class": 3,
                    "visibility": 2,
                    "barcode": "",
                    "dimensions": "",
                    "meta_title": "",
                    "meta_keyword": "",
                    "meta_description": ""
                }
            },
...
            "AddConfigurableRequest": {
                "title": "AddConfigurableRequest",
                "required": [
                    "product_ids"
                ],
                "type": "object",
                "properties": {
                    "product_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "description": ""
                    }
                },
                "example": {
                    "product_ids": [
                        6848,
                        6869
                    ]
                }
            },
            "RemoveConfigurableRequest": {
                "title": "RemoveConfigurableRequest",
                "required": [
                    "product_ids"
                ],
                "type": "object",
                "properties": {
                    "product_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "description": ""
                    }
                },
                "example": {
                    "product_ids": [
                        6848,
                        6869
                    ]
                }
            },
...
            "AssignAttributeRequest": {
                "title": "AssignAttributeRequest",
                "required": [
                    "attribute_ids"
                ],
                "type": "object",
                "properties": {
                    "attribute_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "description": ""
                    }
                },
                "example": {
                    "attribute_ids": [
                        1
                    ]
                }
            },
            "UnassignAttributeRequest": {
                "title": "UnassignAttributeRequest",
                "required": [
                    "attribute_ids"
                ],
                "type": "object",
                "properties": {
                    "attribute_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "description": ""
                    }
                },
                "example": {
                    "attribute_ids": [
                        1
                    ]
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [],
...
}